fix: rename utils folder to mergeClassnames [MDS-343] (#2055)
