+++ workspaces/core/src/avatar/Avatar.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { GenericUser } from '@heathmont/moon-icons-tw';
-import mergeClassnames from '../utils/mergeClassnames';
+import mergeClassnames from '../mergeClassnames/mergeClassnames';
 import setIconSize from './private/utils/setIconSize';
 import Status from './styles/Status';
 import Wrapper from './styles/Wrapper';
