+++ workspaces/core/src/breadcrumb/private/Collapsed.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { ArrowsRight, Other3DotsHorizontal } from '@heathmont/moon-icons-tw';
 import Button from '../../button/Button';
-import mergeClassnames from '../../utils/mergeClassnames';
+import mergeClassnames from '../../mergeClassnames/mergeClassnames';
 import useClickOutside from './hooks/useClickOutside';
 import type BreadcrumbProps from './types/BreadcrumbProps';
