+++ workspaces/core/src/dropdown/Dropdown.tsx
@@ -10,7 +10,7 @@ import React, {
 import { Listbox } from '@headlessui/react';
 import { usePopper } from 'react-popper';
 import { SelectButton } from '../index';
-import mergeClassnames from '../utils/mergeClassnames';
+import mergeClassnames from '../mergeClassnames/mergeClassnames';
 
 type DropdownState = {
   value?: any;
