+++ workspaces/core/src/index.ts
@@ -30,6 +30,8 @@ export { default as Loader } from './loader/Loader';
 export * from './loader/Loader';
 export { default as MenuItem } from './menuItem/MenuItem';
 export * from './menuItem/MenuItem';
+export { default as mergeClassnames } from './mergeClassnames/mergeClassnames';
+export * from './mergeClassnames/mergeClassnames';
 export { default as Modal } from './modal/Modal';
 export * from './modal/Modal';
 export { default as Popover } from './popover/Popover';
@@ -58,5 +60,3 @@ export { default as Textarea } from './textarea/Textarea';
 export * from './textarea/Textarea';
 export { default as Tooltip } from './tooltip/Tooltip';
 export * from './tooltip/Tooltip';
-export { default as mergeClassnames } from './utils/mergeClassnames';
-export * from './utils/mergeClassnames';
