+++ workspaces/core/src/select/Select.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import { Listbox } from '@headlessui/react';
 import { usePopper } from 'react-popper';
-import mergeClassnames from '../utils/mergeClassnames';
+import mergeClassnames from '../mergeClassnames/mergeClassnames';
 import HintText from './private/HintText';
 import InputBtn from './private/InputBtn';
 import Options from './private/Options';
