+++ workspaces/core/src/select/private/InputBtn.tsx
@@ -6,7 +6,7 @@ import {
   ControlsChevronUp,
   ControlsChevronUpSmall,
 } from '@heathmont/moon-icons-tw';
-import mergeClassnames from '../../utils/mergeClassnames';
+import mergeClassnames from '../../mergeClassnames/mergeClassnames';
 
 type InputBtnProps = {
   size?: 'md' | 'lg' | 'xl' | string;
