+++ workspaces/core/src/textInput/private/Container.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import mergeClassnames from '../../utils/mergeClassnames';
+import mergeClassnames from '../../mergeClassnames/mergeClassnames';
 import getBorderRadius from './utils/getBorderRadius';
 
 const Container: React.FC<{ disabled?: boolean; inputSize?: string }> = ({
