+++ workspaces/core/src/textInputGroup/styles/Container.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import mergeClassnames from '../../utils/mergeClassnames';
+import mergeClassnames from '../../mergeClassnames/mergeClassnames';
 
 const Container: React.FC<{
   orientation?: 'vertical' | 'horizontal';
