+++ workspaces/core/src/tooltip/Tooltip.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import * as RadixTooltip from '@radix-ui/react-tooltip';
+import mergeClassnames from '../mergeClassnames/mergeClassnames';
 import useRegisterChild from '../private/utils/useRegisterChild';
-import mergeClassnames from '../utils/mergeClassnames';
 import getAlign from './private/utils/getAlign';
 import getSide from './private/utils/getSide';
 import TooltipContext from './private/utils/TooltipContext';
